requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
lxml>=5.0.0
pytest>=8.0.0
